refactor: 이미지 분석 페이지 - 이미지/디테일정보 컴포넌트 분리
